# docs/swagger.json
{"swagger":"2.0","info":{"description":"サンプルAPIの説明","title":"サンプルAPI"},"host":"localhost","basePath":"/","tags":[{"name":"Sample","description":"サンプル"}],"paths":{"/sample":{"get":{"tags":["Sample"],"summary":"GetSample","description":"Sampleを取得するAPI","operationId":"getSampleUsingGET","produces":["*/*"],"parameters":[{"in":"body","name":"sampleId","description":"サンプルID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","schema":{"originalRef":"GetSampleResponse","$ref":"#/definitions/GetSampleResponse"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["Sample"],"summary":"PostSample","description":"Sampleを登録するAPI","operationId":"postSampleUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"sampleId","in":"query","description":"サンプルID","required":true,"type":"string","x-example":"HOGEHOGE"},{"name":"sampleName","in":"query","description":"サンプル名","required":false,"type":"string","x-example":"サンプル名"},{"name":"sampleType","in":"query","description":"サンプル種別","required":true,"type":"integer","format":"int32","x-example":1}],"responses":{"200":{"description":"OK","schema":{"originalRef":"PostSampleResponse","$ref":"#/definitions/PostSampleResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"GetSampleResponse":{"type":"object","required":["sampleid","sampletype"],"properties":{"sampleid":{"type":"string","example":"HOGEHOGE","description":"サンプルID"},"samplename":{"type":"string","example":"サンプル名","description":"サンプル名"},"sampletype":{"type":"integer","format":"int32","example":1,"description":"サンプル種別"}},"title":"GetSampleResponse"},"PostSampleResponse":{"type":"object","required":["result"],"properties":{"result":{"type":"boolean","example":true,"description":"結果"}},"title":"PostSampleResponse"}}}